Generate Module::CoreList::Data
This commit moves Perl version specific files into a data/ directory
and uses them to generate Module::CoreList::Data, which contains
a compressed dump of corelist data in the __DATA__ section.
Module::CoreList gets an extracted copy of this data on load.

This has the benefit of keeping perl-version files distinct for easy
maintenance, but significantly shrinks the installed size of the
data.